Refactor pipeline stages for suggestions
